app/requirements.txt
streamlit>=1.30.0
streamlit-agraph>=0.0.45
networkx>=3.0
openai>=1.0.0
databricks-sdk>=0.18.0
chromadb>=0.4.0
